move run to pkg/server
