+++ pkg/ebpf/ebpf.go
@@ -2,7 +2,6 @@ package ebpf
 
 import (
 	"bytes"
-	"context"
 	"encoding/binary"
 	"errors"
 	"log"
@@ -11,8 +10,6 @@ import (
 	"github.com/cilium/ebpf/link"
 	"github.com/cilium/ebpf/ringbuf"
 	"github.com/cilium/ebpf/rlimit"
-	awscloud "github.com/jrroman/caza/pkg/cloud/aws"
-	"github.com/jrroman/caza/pkg/config"
 	"github.com/jrroman/caza/pkg/metrics"
 	"github.com/prometheus/client_golang/prometheus"
 )
@@ -110,7 +107,7 @@ func reportNetworkProximity(pair *NetworkPair) string {
 
 // We need to process the events being sent down the event channel. As a first stab
 // lets create an IP map it could look something like map[net.IP]struct{in out}
-func processEvents(ctx context.Context, ec chan bpfEvent, networks map[string]*net.IPNet) {
+func ProcessEvents(ec chan *NetworkPair, networks map[string]*net.IPNet) {
 	log.Printf("Processing events for networks:\n%v", networks)
 	log.Printf("%-15s %-6s -> %-15s %-6s %-6s",
 		"Src addr",
@@ -119,20 +116,17 @@ func processEvents(ctx context.Context, ec chan bpfEvent, networks map[string]*n
 		"Port",
 		"Type",
 	)
-	for event := range ec {
-		srcAddr, dstAddr := intToIP(event.Saddr), intToIP(event.Daddr)
-		srcPort, dstPort := event.Sport, event.Dport
-		pair := createNetworkPairRelation(
-			networks, newNetworkPair(srcAddr, dstAddr, srcPort, dstPort))
+	for eventData := range ec {
+		pair := createNetworkPairRelation(networks, eventData)
 		proximity := reportNetworkProximity(pair)
 		log.Printf("%-15s %-6d -> %-15s %-6d %-6s",
-			srcAddr, srcPort, dstAddr, dstPort, proximity)
+			pair.src.ip, pair.src.port, pair.dst.ip, pair.dst.port, proximity)
 	}
 }
 
 // We need to read the events being sent from our fentry.c program via ringbuffer.
 // These events occur on tcp_close events and contain the src and dest ip and port.
-func eBPFReadLoop(rd *ringbuf.Reader, ec chan bpfEvent) {
+func eBPFReadLoop(rd *ringbuf.Reader, ec chan *NetworkPair) {
 	// bpfEvent is generated by bpf2go.
 	var event bpfEvent
 	for {
@@ -150,13 +144,13 @@ func eBPFReadLoop(rd *ringbuf.Reader, ec chan bpfEvent) {
 			log.Printf("parsing ringbuf event: %s", err)
 			continue
 		}
-		ec <- event
+		srcAddr, dstAddr := intToIP(event.Saddr), intToIP(event.Daddr)
+		srcPort, dstPort := event.Sport, event.Dport
+		ec <- newNetworkPair(srcAddr, dstAddr, srcPort, dstPort)
 	}
 }
 
-func eBPFRun(ctx context.Context, ec chan bpfEvent) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
+func LoadAndRun(ec chan *NetworkPair) {
 	// Allow current process to lock memory for eBPF resources
 	if err := rlimit.RemoveMemlock(); err != nil {
 		log.Printf("Remove memory lock: %v", err)
@@ -193,29 +187,4 @@ func eBPFRun(ctx context.Context, ec chan bpfEvent) {
 	}
 	defer rd.Close()
 	eBPFReadLoop(rd, ec)
-}
-
-func Run(ctx context.Context, cfg *config.Config) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	var networkSlice []map[string]*net.IPNet
-	if len(cfg.Networks) != 0 {
-		networkSlice = append(networkSlice, cfg.Networks)
-	}
-	if cfg.CloudEnabled {
-		awscc, err := awscloud.New(cfg.Region)
-		if err != nil {
-			return err
-		}
-		cloudNetworks, err := awscc.GetNetworks(cfg.VpcID)
-		if err != nil {
-			return err
-		}
-		networkSlice = append(networkSlice, cloudNetworks)
-	}
-	eventChan := make(chan bpfEvent)
-	go eBPFRun(ctx, eventChan)
-	go processEvents(ctx, eventChan, mergeNetworkMaps(networkSlice))
-	<-ctx.Done()
-	return nil
 }
