+++ pkg/ebpf/helpers.go
@@ -31,19 +31,3 @@ func intToIP(ipNum uint32) net.IP {
 	NativeEndian.PutUint32(ip, ipNum)
 	return ip
 }
-
-// TODO pull this network data in from aws or whatever cloud provider we are utilizing
-// This will end up being in its own package so we can utilize the aws api
-func mergeNetworkMaps(networks []map[string]*net.IPNet) map[string]*net.IPNet {
-	// if there is only one network return it
-	if len(networks) == 1 {
-		return networks[0]
-	}
-	merged := make(map[string]*net.IPNet)
-	for _, nm := range networks {
-		for name, network := range nm {
-			merged[name] = network
-		}
-	}
-	return merged
-}
